Changed db columns to big int

+++ database/migrations/2018_01_31_221341_create_mothers_table.php
@@ -14,12 +14,12 @@ class CreateMothersTable extends Migration
     public function up()
     {
         Schema::create('mothers', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('national_mother_id')->unsigned()->nullable();
+            $table->bigIncrements('id');
+            $table->bigInteger('national_mother_id')->unsigned()->nullable();
             $table->string('ccc_no', 50)->nullable()->index();
             // $table->integer('fcode')->unsigned()->nullable();
             $table->integer('facility_id')->unsigned()->index();
-            $table->integer('hiv_status')->unsigned();
+            $table->tinyInteger('hiv_status')->unsigned();
             $table->tinyInteger('synched')->default(0);
             $table->date('datesynched')->nullable();
             // $table->date('created_at')->nullable();
